feat: delete signin page, delete console logs
