+++ features/auth/AuthIcons.tsx
@@ -2,27 +2,59 @@ import { Path } from 'common/enums/path'
 import Image from 'next/image'
 import { useRouter } from 'next/router'
 import styled from 'styled-components'
-
+import { ProvidersPropsType } from 'features/auth/types'
 import github from '../../public/img/icons/github-svgrepo-com.svg'
 import google from '../../public/img/icons/google-svgrepo-com.svg'
 import { baseTheme } from '../../styles/styledComponents/theme'
+import { oauthRequest } from './oauth2Request'
+import { GetStaticPropsContext } from 'next'
+import { useTranslation } from 'next-i18next'
 
-const AuthIcons = () => {
+const AuthIcons = (providerParams: ProvidersPropsType) => {
   const { push } = useRouter()
+  const { t } = useTranslation()
+
+  const handle = (providerName: string) => {
+    const url = oauthRequest(providerName, providerParams)
+
+    console.log(url)
+    window.location.assign(url)
+  }
 
   return (
     <StyledIconBlock>
-      <Link
+      {/* <Link
         href="/api/auth/signin"
         onClick={async e => {
           e.preventDefault()
           push(Path.SIGNIN)
         }}
-      >
-        <Image alt="google" height={36} src={google} width={36} />
-        <Image alt="github" height={36} src={github} width={36} />
-      </Link>
-      <Message>SignIn with Google or Github</Message>
+      > */}
+      <SigninStyle>
+        <Image
+          onClick={() => {
+            handle('google')
+          }}
+          alt="google"
+          src={google}
+          height={36}
+          width={36}
+        />
+      </SigninStyle>
+
+      <SigninStyle>
+        <Image
+          onClick={() => {
+            handle('github')
+          }}
+          alt="github"
+          src={github}
+          height={36}
+          width={36}
+        />
+      </SigninStyle>
+
+      {/* </Link> */}
     </StyledIconBlock>
   )
 }
@@ -45,6 +77,9 @@ const StyledIconBlock = styled.div`
     display: block;
     text-align: center;
   }
+`
+const SigninStyle = styled.div`
+  cursor: pointer;
 `
 
 const Link = styled.a`
